remove CommitLogElementsInsert replace EnsureCategoryAndPortExist with SaveChangesToPortsTree create _CompileListOfPorts which replaces the fuctionality of EnsureCategoryAndPortExist
Instead of saving a port as we go along, we will compile a list of ports based
on the files we find.  Once all files are processed and saved to the database,
we have our list.  Then we process the files again, setting the needs_refresh
value for the port affected by this file.  If we find that the Makefile for
a port has been deleted, we set needs_refresh = 0.


git-svn-id: svn+ssh://svn.int.unixathome.org/usr/local/svn/repos/freshports-1@654 46e82423-29d8-e211-989e-002590a4cdd4