Add missing.php.
This page is activated via ErrorDocument 404 /missing.php in the apache
configuration for this website.  It uses the classes element and port to
grab information from the database.

This is the start of being able to browse to:

http://FreshPorts.org/net/samba and get the samba page.  Or, if a port
is uniqely named, http://FreshPorts.org/samba will take you to the
same place.

Much to do, but this is the proof of concept.


git-svn-id: svn+ssh://svn.int.unixathome.org/usr/local/svn/repos/freshports-1@743 46e82423-29d8-e211-989e-002590a4cdd4