upgrade to postgresql.
Now that this is done, the next step is to get some of you on the Develop list
to start tracking your ports via fp2.freshports.org.  BUT FIRST I'll have to
upgrade that site to the latest database and scripts as found in
lists-test.freshports.org.  Watch this space.


git-svn-id: svn+ssh://svn.int.unixathome.org/usr/local/svn/repos/freshports-1@824 46e82423-29d8-e211-989e-002590a4cdd4