A port always has one primary category.  This category corresponds to the directory in which the port resides. This directory is always within the ports tree.  A port may also reside in other categories but they are referred to as secondary categories.
When displaying the contents of a category, the same port name may appear
twice, but from different categories.  For example, japanese/gawk and
lang/gawk.  The query to obtain a list of ports from a given category
will now return both the primary category (as category) and the name
of the category we are looking at (as category_looking_at).


git-svn-id: svn+ssh://svn.int.unixathome.org/usr/local/svn/repos/freshports-1@2399 46e82423-29d8-e211-989e-002590a4cdd4