We aren't using jquery-1.5.min.js
Remove it

The packageToolTipText was displaying the import date instead of the processed date.

Fix that.

The date('Y-m-d H:i', strtotime($import_date)) invocation was converting to another timezone.
We don't want that.  We want UTC. Instead, let the database supply the correctly formatted date.
We just display it.