port-display.php: id masterport and id slaveports
Add id `masterport` to the ***Master port**:* line. (Is `class="masterport"` required? I assume so.)

Move id `slaveports` up to the _Slave ports:_ heading for the ordered list, instead of the first numbered item within the list. (Is this valid? Recalling https://github.com/FreshPorts/freshports/issues/196#issuecomment-766402756 where not everything that  _appears_ to be an anchor is intended for page positioning/linking.)