Fix links to svn repo
The links to the subversion repo are wrong because
freshports_svnweb_ChangeSet_Link() was being invoked incorrectly.  Too many
parameters.

Simlarly, freshports_svnweb_ChangeSet_Link_Text() had a unused $part argument -
removed.

While there, the $path argument is not used: remove it.

In the commit history for a port:

* instead of showing the short hash for git commits, display
  the git icon.
* put the git icon or svn icon before the PORTVERSION - this should make the
  date column just one line

Fixes #242