Order by CL.message_id
We once ordered by 1 desc, commit_log.id

But that would give different results between dev and test if a commit had to
be rerun on dev. In that case, that commit would get a a higher commit id than
commits which preceeded it.

Why sort by mesage_id at all? Keep together all ports for a given commit keep.