Use port commits only.
The results say 'commits', it should say 'ports'.

The next decision: if we are searching commit messages, do we show ports in the
results? We could show commits, then you click on the commit to see the ports
(if any) affected by that commit. This would make it easier to also include src
commits, if that option is selected.

Fixes #431 sort of.