Recognize and display older Subversion commits
Checking for freebsd.org in the message id harks back to subversion days.
That's when FreshPorts received and parsed emails from the mailing lists.
Message-ID was taken from the email headers.
Thus, this if is checking to see if this is a subversion email, and if not, assume it is a git commit.
However, there is a another class of commits which did not store message_id.
When that column was added, existing commits were given a value similar to: fp1.9826@dev.null.freshports.org
This if says: if not a subversion commit, display the git commit.

That is, if the message_id field  contains 'freebsd.org' or 'dev.null.freshports.org, it's subversion.