REQUEST_URI is not a URL.
Instead, it is the bit after the hostname e.g. /index.html

re: https://www.php.net/manual/en/reserved.variables.server.php

What we need for parse_url is a URL, with schema, hostname, etc.

So start creating that ourselves via:

$url = $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];

We invoke parse_url to grab the QUERY:

$url_query = parse_url($url, PHP_URL_QUERY);

Previously, parse_url() was invoked on REQUEST_URI, but that interprets: //d as
a hostname instead of
path.

array(1) {
  ["host"]=>
  string(1) "d"
}

If we use a whole URL (e.g. https://www.freshports.org//d) in the call to
parse_url(), we get the expected results.

array(3) {
  ["scheme"]=>
  string(5) "https"
  ["host"]=>
  string(18) "www.freshports.org"
  ["path"]=>
  string(3) "//d"
}