classes/port-display.php: Better handling for ports not on quarterly
The current version (PORTVERSION) of a port on head is important.
So is the version on quarterly.

The code grabs this from the database as it's being fetched.

If there's a commit on quarterly, grab that. If not, grab the
latest commit before the quarter was created.

If nothing found, the port was created during the current quarter.
If that's the case, let the user know.