Correct the URLs for github and gitlab
Previously, these functions assumed the commit was in the ports tree. Now they
use the incoming $path variable. This value is taken from the FreshPorts
database and reflect the location with the source tree as stored in FreshPorts.

The code translates this value to the values used by GitHub and GitLab
respectively.

This fixes #392